+++ src/main/java/paintingcanvas/animation/RotationAnimation.java
@@ -19,15 +19,15 @@ public class RotationAnimation extends Animation {
         relative = true;
         return this;
     }
-    
+
     @Override
     public Animation copy() {
         return new RotationAnimation(end);
     }
-    
+
     @Override
     protected void updateAnimation(Drawable drawable, double progress) {
-        var t = (double) easing.ease(progress);
+        var t = easing.ease(progress);
         drawable.rotation = start + (end - start) * t;
     }
